Bill PR
PR includes the following WIP changes: 

- Adds very basic ragequit for whitelisted tokens in the contract
- Uses ragequit for member kicks 
- Splits out proposal processing into different functions based on type of proposal
- Adds early execution as a possibility for proposals 
- Adds ability to cancel proposals that haven't been processed 

Also, stuck with the bool for flags after realizing that this will probably allow for it to be used with existing minions. 

Open Question: 
With the additional of ragequit should we introduce some sort of grace period concept again? I think this might make sense, otherwise RQ only works in the scenario where the quitter votes no on a proposal and thinks it's going to pass. 

NOTE: Haven't tested all of these changes, but want to get them into a PR for conceptual review.