Store Calldata hash instead of raw calldata bytestring
it's much cheaper and less variable to store a bytes32 representation of
the calldata rather than the calldata itself on chain. The calldata to
be processed after a proposal passes can be pulled from event storage
instead, and passed back into the `processProposal` function.

As a sanity check, the hash of the calldata is emitted in the
`submitProposal` event. This mimics the moloch v2 behavior but can be
removed if we want to optimize event-level storage as well.

test plan:
unit tests updated to verify that mismatched calldata passed into
`processProposal` will yield a revert. There doesn't appear to be much
testing of data emitted from events, so that path was skipped for now.